Let authors pick interactive for the image question
[#158702375]

+ specs
+ refactoring of common functionality (shared between labbooks
  and image questions)
