+++ app/models/embeddable/image_question.rb
@@ -1,8 +1,9 @@
 class Embeddable::ImageQuestion < ActiveRecord::Base
+  include Embeddable
+  include AttachedToInteractive
+
   attr_accessible :name, :prompt, :hint, :bg_source, :bg_url, :drawing_prompt, :is_full_width,
     :is_prediction, :is_featured, :give_prediction_feedback, :prediction_feedback, :is_hidden
-
-  include Embeddable
 
   has_many :page_items, :as => :embeddable, :dependent => :destroy
   has_many :interactive_pages, :through => :page_items
@@ -17,14 +18,6 @@ class Embeddable::ImageQuestion < ActiveRecord::Base
   has_one :master_for_tracker, :class_name => 'QuestionTracker', :as => :master_question
 
   default_value_for :prompt, "why does ..."
-
-  def interactive
-    # Return first interactive available on the page (note that in practice it's impossible that this model has more
-    # than one page, even though it's many-to-many association).
-    # In the future we can let authors explicitly select which interactive an image question is connected to.
-    page = interactive_pages.first
-    page && page.visible_interactives.first
-  end
 
   # NOTE: publishing to portal doesn't use this hash. portal_hash is used instead
   def to_hash
@@ -97,9 +90,20 @@ class Embeddable::ImageQuestion < ActiveRecord::Base
     true
   end
 
+  def page
+    # Return first page (note that in practice it's impossible that this model has more
+    # than one page, even though it's many-to-many association).
+    interactive_pages.first
+  end
+
   def page_section
     # In practice one question can't be added to multiple pages. Perhaps it should be refactored to has_one / belongs_to relation.
     page_items.count > 0 && page_items.first.section
+  end
+
+  def configuration_error
+    return I18n.t('SNAPSHOT_WITHOUT_INTERACTIVE') if is_shutterbug? && interactive.nil?
+    nil
   end
 
   def self.name_as_param
