+++ spec/models/embeddable/image_question_spec.rb
@@ -2,6 +2,7 @@ require 'spec_helper'
 
 describe Embeddable::ImageQuestion do
   it_behaves_like "a question"
+  it_behaves_like "attached to interactive"
 
   let (:image_question) { FactoryGirl.create(:image_question) }
