+++ spec/support/shared_examples/attached_to_interactive_form.rb
@@ -1,0 +1,38 @@
+shared_examples "attached to interactive form" do
+  def interactive_choice_select_css
+    "li.interactive_choice select option"
+  end
+
+  def no_interactive_text
+    AttachedToInteractive::NO_INTERACTIVE_LABEL
+  end
+
+  let(:interactive_a)  { FactoryGirl.create(:mw_interactive) }
+  let(:interactive_b)  { FactoryGirl.create(:mw_interactive) }
+  let(:interactives)   { [] }
+  let(:page)           { FactoryGirl.create(:page_with_or, interactives: interactives) }
+  let(:args)           { {} }
+
+  before :each do
+    page.add_embeddable(test_embeddable)
+  end
+
+  describe "without any interactives on the page" do
+    it "includes one choice for 'no interactive'" do
+      assign(:embeddable, test_embeddable)
+      render
+      expect(rendered).to have_css(
+        interactive_choice_select_css,
+        text: no_interactive_text)
+    end
+  end
+
+  describe "with two interactives on the page" do
+    let(:interactives)  {[interactive_a,interactive_b]}
+    it "includes three choices" do
+      assign(:embeddable, test_embeddable)
+      render
+      expect(rendered).to have_css interactive_choice_select_css, count: 3
+    end
+  end
+end
